backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
azure-core==1.29.5
azure-identity==1.15.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
distro==1.8.0
exceptiongroup==1.2.0
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.1
google-search-results==2.4.1
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.2
keyring==24.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.1.0
msal==1.26.0
msal-extensions==1.1.0
nltk==3.8.1
numpy==1.26.2
openai==1.3.8
opencensus==0.11.3
opencensus-context==0.1.3
opencensus-ext-azure==1.1.12
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
portalocker==2.8.2
promptflow==1.1.1
promptflow-tools==1.0.0
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydash==7.0.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
strictyaml==1.7.3
tabulate==0.9.0
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
waitress==2.1.2
Werkzeug==3.0.1
zipp==3.17.0
faiss-cpu==1.7.4